test (TestFile) : 가독성 향상을 위한 테스트 클래스 메소드 분리 및 상수화
