+++ docs/README.md
@@ -64,7 +64,6 @@
 - [x] 입력 형식의 유효성 검증
   - [x] 메뉴-갯수 조합 모임 입력 형식의 유효성 검증
     - [x] 공백이 아닌지
-    - ~~[ ] ","로 구분되어 들어오는지~~
     - [x] ","가 2번이상 연속으로 들어가지 않는지
     - [x] 맨 앞, 뒤에 ","가 있지 않은지
   - [x] 개별 메뉴-갯수 조합 입력 형식의 유효성 검증 
@@ -103,11 +102,10 @@
   - [x] 메소드 순서
 - [x] 이해를 위한 코드 전반의 공백 추가
 - [x] 메서드 길이가 15자를 넘어가지 않는지 확인
-- [ ] 필요한 테스트가 되어있는지 확인
-- [ ] 테스트 클래스 리팩토링
+- [x] 필요한 테스트가 되어있는지 확인
+- [x] 테스트 클래스 리팩토링
 - [ ] 의존관계 그려보고 조정하기
 - [ ] 요구사항 하나씩 줄쳐가면서 빠뜨린 것 없는지 확인
-- [ ] 각 separator의 위치 고민
 - [x] DateGroup 방식 간소화 또는 enum화
 - [x] freeGift NOTHING(Menu.NOTHING, 0) 으로 인한 메뉴 NOTHING 
 - [x] "타파스-1,제로콜라-1 " 처럼 앞 뒤 공백 예외처리 안되게
