+++ src/test/java/christmas/constant/testConstant.java
@@ -1,0 +1,10 @@
+package christmas.constant;
+
+public class testConstant {
+
+    public static final int WEEKEND_D_DAY1 = 1;
+    public static final int WEEKDAY_SPECIAL_D_DAY3 = 3;
+    public static final int WEEKDAY_SPECIAL_D_DAY25 = 25;
+    public static final int WEEKDAY_DAY26 = 26;
+
+}
