+++ src/test/java/christmas/domain/MenuCountTest.java
@@ -1,20 +1,19 @@
-package christmas;
+package christmas.domain;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import christmas.common.ErrorMessage;
-import christmas.domain.VisitDate;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-public class VisitDateTest {
-    @DisplayName("날짜는 1~31일 사이여야 한다.")
+public class MenuCountTest {
+    @DisplayName("개별 메뉴 개수는 1보다 커야한다.")
     @ParameterizedTest
-    @ValueSource(ints = {-1, 0, 32})
-    void inputDay(int input) {
+    @ValueSource(ints = {-1, 0})
+    void validateMenuCount(int input) {
         assertThatIllegalArgumentException()
-                .isThrownBy(() -> VisitDate.from(input))
-                .withMessage(ErrorMessage.DATE_IS_INVALID);
+                .isThrownBy(() -> MenuCount.from(input))
+                .withMessage(ErrorMessage.ORDER_IS_INVALID);
     }
 }
