+++ src/test/java/christmas/domain/MenuTest.java
@@ -1,10 +1,9 @@
-package christmas;
+package christmas.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
 import christmas.common.ErrorMessage;
-import christmas.domain.Menu;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
@@ -19,6 +18,7 @@ public class MenuTest {
                 .isThrownBy(() -> Menu.from(input))
                 .withMessage(ErrorMessage.ORDER_IS_INVALID);
     }
+
     @DisplayName("메뉴판에 있는 메뉴면 이름에 맞는 메뉴를 반환한다.")
     @ParameterizedTest
     @CsvSource(value = {"양송이수프, MUSHROOM_SOUP"})
