--- src/test/java/christmas/testConstant.java
@@ -1,4 +1,0 @@
-package christmas;
-
-public class testConstant {
-}
